#9 グラフィック系のファイル名の変更
