+++ ProjectDxLibEx/Graph2D.cpp
@@ -5,7 +5,7 @@
   Distributed under the Boost Software License, Version 1.0.
   (See http://www.boost.org/LICENSE_1_0.txt)
 =============================================================================*/
-#include "dxlibex/graph2d.h"
+#include "dxlibex/graph2d.hpp"
 
 namespace dxle{
 namespace graph2d {
