+++ dxlibex/config/no_min_max.h
@@ -5,8 +5,8 @@
   Distributed under the Boost Software License, Version 1.0.
   (See http://www.boost.org/LICENSE_1_0.txt)
 =============================================================================*/
-#ifndef DXLE_INC_CONFIG_NO_MIN_MAX_HPP_
-#define DXLE_INC_CONFIG_NO_MIN_MAX_HPP_
+#ifndef DXLE_INC_CONFIG_NO_MIN_MAX_H_
+#define DXLE_INC_CONFIG_NO_MIN_MAX_H_
 
 //min max É}ÉNÉçñ≥å¯âª
