+++ dxlibex/texture/animation_graph.hpp
@@ -5,8 +5,8 @@
   Distributed under the Boost Software License, Version 1.0.
   (See http://www.boost.org/LICENSE_1_0.txt)
 =============================================================================*/
-#ifndef DXLE_INC_TEXTURE_ANIMATION_GRAPH_H_
-#define DXLE_INC_TEXTURE_ANIMATION_GRAPH_H_
+#ifndef DXLE_INC_TEXTURE_ANIMATION_GRAPH_HPP_
+#define DXLE_INC_TEXTURE_ANIMATION_GRAPH_HPP_
 
 #include "dxlibex/config/no_min_max.h"
 #include <iterator>
@@ -16,7 +16,7 @@
 #include "DxLib.h"
 #include "texture2d.h"
 #include "dxlibex/time.hpp"
-#include "dxlibex/thread.h"
+#include "dxlibex/thread.hpp"
 
 namespace dxle
 {
