+++ dxlibex/texture/texture2d/prototype2d.hpp
@@ -5,8 +5,8 @@
   Distributed under the Boost Software License, Version 1.0.
   (See http://www.boost.org/LICENSE_1_0.txt)
 =============================================================================*/
-#ifndef DXLE_INC_TEXTURE_PROTOTYPE2D_H_
-#define DXLE_INC_TEXTURE_PROTOTYPE2D_H_
+#ifndef DXLE_INC_TEXTURE_TEXTURE2D_PROTOTYPE2D_HPP_
+#define DXLE_INC_TEXTURE_TEXTURE2D_PROTOTYPE2D_HPP_
 
 #include "dxlibex/config/no_min_max.h"
 #include "DxLib.h"
